requirements.txt
alembic==1.3.1
arrow==0.15.4
certifi==2019.9.11
chardet==3.0.4
Click==7.0
dateparser==0.7.2
DateTime==4.3
Flask==1.1.1
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
moment==0.8.2
psycopg2-binary==2.8.4
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.3
regex==2019.11.1
requests==2.22.0
six==1.13.0
SQLAlchemy==1.3.11
times==0.7
tzlocal==2.0.0
urllib3==1.25.7
Werkzeug==0.16.0
zope.interface==4.7.1
